requirements.txt
# ==================== requirements.txt ====================
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
json5>=0.9.0
tqdm>=4.62.0
albumentations>=1.0.0
pycocotools>=2.0.0
detectron2>=0.6.0
transformers>=4.20.0
# For Raspberry Pi camera
picamera2>=0.3.0
# For 3D visualization
open3d>=0.13.0
trimesh>=3.9.0